Alterted room grill column

+++ Back-End/EServices.WebApi/Domain/Entities/RoomGrill.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
     public class RoomGrill : AuditableBaseEntity
     {
         public string ReferenceNumber { get; set; }
-        public int FilterAreaSqft { get; set; }
+        public float FilterAreaSqft { get; set; }
         public string AirVelocityReadingInFPMO { get; set; }
         public int AvgVelocityFPM { get; set; }
         public int AirFlowCFM { get; set; }
